+++ apps/user-manager/src/test/java/com/warlock/user/controller/UserControllerProfileTest.java
@@ -1,0 +1,13 @@
+package com.warlock.user.controller;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+
+public class UserControllerProfileTest extends UserControllerTest{
+    @Test
+    void profile_ReturnsOk() {
+        assertTrue(true);
+    }
+}